updated code after dryrun
